move initial file id
